backend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.1.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
Pillow==6.2.1
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3